refactor: densify great-power war count and military-plans pins (#4310)
Split two remaining ≥450 ungated planning pins into thin contracts plus
sibling case modules; register both on residual-fat gated basenames.

Refs #4310
