+++ packages/colonizethis_ai/test/planning/expand_phase_planner_great_power_war_count_and_pending_declare_war_cases.dart
@@ -1,0 +1,10 @@
+// Case-library barrel for great-power war count / pending declare-war pin
+// (Refs #4310 Slice D).
+
+import 'expand_phase_planner_great_power_war_count_cases.dart';
+import 'expand_phase_planner_pending_declare_war_cases.dart';
+
+void registerExpandPhasePlannerGreatPowerWarCountAndPendingDeclareWarCases() {
+  registerExpandPhasePlannerGreatPowerWarCountCases();
+  registerExpandPhasePlannerPendingDeclareWarCases();
+}
